layout: Tweak top/bottom bars for Venu 3
Tweak the positions and offsets of the top/bottom bars, to
work better on devices with round screens. The top bar was
a bit too short, causing the time to be cut off, while the
bottom was seemed taller than necessary, given the content
that was in it.

This looks okay for me, but may be suboptimal on other
deviecs. Ideally, we would use things like font size to
determine the vertical offset of the time, for example.

Signed-off-by: evilwombat <evilwombat@peregrine.local>