util: suncalc: Fix loss of precision on simulator
It seems like constructing a Moment having fractional
seconds *is* possible, but only if we're running on the
simulator. This results in Moments that are internally
seemingly using a single-precision float to store their
epoch time, which leads to loss of precision when working
with moments having an epoch time of sufficiently high
magnitude.

Signed-off-by: evilwombat <evilwombat@peregrine.local>