ui: Don't draw central ring if not needed
If the daylight indicator is off, don't draw the circle
around the HR indicator. This looks a bit cluttered, but
we may add an option to bring this back, to stay true to
the original UI.

Split the central vertical line into two, to avoid drawing
this line directly through the HR indicator if the central
ring is hidden.

Signed-off-by: evilwombat <evilwombat@peregrine.local>