core: Effectively inline formatInfo()
Rather than stashing all activity information in member
variables and then formatting them on-demand using a large
function, perform all the necessary formatting as part of
compute() and stash the results into an array indexed by
data type. This allows the rendering code to immediately
look up the info string for any field type. Despite needing
to format the strings for each field (regardless of whether
that field is shown), this leads to a net memory savings,
due to reduced code size. The tradeoff is increaed compute
demand for formatting things we may not need.

Signed-off-by: evilwombat <evilwombat@peregrine.local>