settings: Get rid of settings groups :(
Although Groups are nice for organizing app settings, they
seem to be broken in the ConnectIQ app for iOS. So, move
all the settings to a single screen, to work around an iOS
bug where Groups cause the 'Save' button to disappear in
the ConnectIQ app. I may bring these back, if Garmin fixes
this problem.

Signed-off-by: evilwombat <evilwombat@peregrine.local>