core: Try to fix TypeError during goal animation display?
People seem to be hitting crashes in onTimerStop() with a
type error on the following line:
Storage.setValue("startTime",
            Activity.getActivityInfo().startTime.value());

Although this seems to work in the nominal case (starting
and stopping the activity timer manually), today I hit this
seemingly after hitting either a step or activity goal
animation (though I wasn't paying close attention during
the activity).

ERA seems to indicate a number of crashes on this line, so
add a bunch of NULL checks until I can reproduce this more
reliably.

Signed-off-by: evilwombat <evilwombat@peregrine.local>