core: Use distance units for elevation on vivoactive5
The vivoactive 5 system menu seemingly provides no way to
configure the system's elevation units. On this device, use
the distance units setting to determine which elevation
units to use. Do this by splitting the units loading into
a separate function, and add a dedicated version of this
function for vivoactive 5 (and for all others) via an
exclude annotation.

Signed-off-by: evilwombat <evilwombat@peregrine.local>