res: Refactor how mid/short labels are handled
In the past, the default data labels were stored in the
top-level strings.xml file, with certain watch models
gaining a resource override, pointing them at a string.xml
dedicated to watches with mid- or short-length labels. This
is all well and good, except it breaks our ability to do
localization, since Garmin only supports resource language
suffixes on paths on the basis of model or general screen
shape, but not based on an arbitrary list of devices.

To avoid massive amounts of duplication, create language-
specific string resources in the midlabels and shortlabels
directories, and add a Python script that expands these
into a large pile of resource overrides, based on device
model and language.

Signed-off-by: evilwombat <evilwombat@peregrine.local>