core: Don't cache common layout variables
Don't cache centerX and centerY as member variables, since
accessing these takes a fair amount of code. Retrieve these
each time we need them, which ends up being a net improvement
in memory usage. Similarly, cache the Y coordinate of the
center circle, so we don't have to calculate it in multiple
places.

Signed-off-by: evilwombat <2313235+evilwombat@users.noreply.github.com>