build: Add framework for autogenerated per-device constants
The only per-device build flags supported by the jungle
build system seem to be exclusion annotations. Although
these can "in theory" be used to implement different per-
device behaviors at compile time, these come at a cost, in
that we can only exclude a function at build time, but we
must still pay the memory penalty associated with defining
an empty function. Furthermore, in some situations, we may
need to define a per-device constant having more than two
possible values/behaviors, meaning we need to add an
exclude annotation _and its negation_ for every variant.

That said, the jungle build system *does* allow us to
define per-device source paths. But rather than replicating
the main source file across all devices, define a per-
device file of *constants*. This way, each device build
will use the device-specific constant definitions when
building the (shared) main source file. The per-device
constants files can be generated from a common device
table, so that all constants are visible in one place.

This is "sort of" similar to per-device properties, but
with the advantage that the values of the per-device
constants are known at build time, and can be used in
compiler optimizations. This also gives us the crude
abilty to exclude certain code blocks from certain devices,
by wrapping them in an if (CONST) {} block, where CONST
is statically defined to be true or false, on a per-device
basis. This isn't as liberating as using #ifdefs, as the
code blocks being excluded have to be syntactically valid
in all situations (ie, we can't ifdef out a variable
definition), but it's better than nothing.

Signed-off-by: evilwombat <2313235+evilwombat@users.noreply.github.com>