core: Handle elevation usings through per-device constants
Previously, we used exclude annotations to indicate whether
a device had a user-facing elevation units setting. Although
this works, this adds a lot of boilerplate to the jungle
file, and has a little bit of extra overhead due to needing
a function call that can be selectively excluded.

Instead, indicate the presence of elevation units using a
per-device constant, so that the code paths for loading
these units can be optimized in/out at build time, without
having to explicitly make a function call.

Signed-off-by: evilwombat <2313235+evilwombat@users.noreply.github.com>